% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cab_drugs.R
\name{cab_drugs}
\alias{cab_drugs}
\title{get drugs records from different places in caboode}
\usage{
cab_drugs(
  con,
  tablefrom = "events",
  patients = "out",
  dayfrom = NULL,
  dayto = NULL,
  name_part = "POSACONAZOLE",
  mrns = NULL,
  collect = TRUE
)
}
\arguments{
\item{con}{database connection to caboodle}

\item{tablefrom}{'orders', 'events', 'components'}

\item{patients}{'in', 'out', 'both'}

\item{dayfrom}{optional 20230601 #caboodle date format}

\item{dayto}{optional 20230602}

\item{name_part}{e.g. 'POSACONAZOLE'}

\item{mrns}{optional patient mrns}

\item{collect}{whether to collect results from database at end}
}
\description{
get drugs records from different places in caboode
}
