% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cab_drugs.R
\name{cab_drugs_events}
\alias{cab_drugs_events}
\title{get drugs records from events table in caboode}
\usage{
cab_drugs_events(
  con,
  patients = "out",
  dayfrom = NULL,
  dayto = NULL,
  name_part = "POSACONAZOLE",
  mrns = NULL,
  collect = TRUE
)
}
\arguments{
\item{con}{connection to caboodle (or maybe clarity too later)}

\item{patients}{'in', 'out', 'both'}

\item{dayfrom}{optional 20230601 #caboodle date format}

\item{dayto}{optional 20230602}

\item{name_part}{e.g. 'POSACONAZOLE'}

\item{mrns}{optional patient mrns}

\item{collect}{whether to collect results from database at end}
}
\description{
get drugs records from events table in caboode
}
\examples{
# example code
}
