% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cab_mrn.R
\name{cab_mrn}
\alias{cab_mrn}
\title{join PrimaryMrn from Caboodle onto passed dataframe}
\usage{
cab_mrn(df, con, name_dk = "DurableKey", name_mrn = "PrimaryMrn")
}
\arguments{
\item{df}{dataframe containing a durablekey column}

\item{name_dk}{optionally specify durable key name if it is not "DurableKey"}

\item{name_mrn}{optionally specify mrn column name in output if not "PrimaryMrn"}
}
\description{
join PrimaryMrn from Caboodle onto passed dataframe
}
\examples{
#cab_mrn(con,name_dk="PatientDurableKey,name_mrn="PatientMrn")
}
