% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caboodler-package.R
\docType{package}
\name{caboodler-package}
\alias{caboodler}
\alias{caboodler-package}
\title{caboodler: Eases access to Caboodle data warehouse from Epic Electronic Health Record}
\description{
For querying Electronic Health Record data in Epic Caboodle data warehouse.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andysouth/caboodler}
  \item Report bugs at \url{https://github.com/andysouth/caboodler/issues}
}

}
\author{
\strong{Maintainer}: Andy South \email{southandy@gmail.com} (\href{https://orcid.org/0000-0003-4051-6135}{ORCID})

}
\keyword{internal}
